# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ExportPriceConfigureRequest": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "amount_minor": {
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "enum": [
              "INR",
              "USD"
            ],
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "currency",
          "amount_minor"
        ],
        "title": "ExportPriceConfigureRequest",
        "type": "object"
      },
      "ExportPriceResponse": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "enum": [
              "INR",
              "USD"
            ],
            "title": "Currency",
            "type": "string"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "currency",
          "amount_minor",
          "active",
          "version",
          "effective_at",
          "created_at"
        ],
        "title": "ExportPriceResponse",
        "type": "object"
      },
      "ExportPurchaseResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "price": {
            "$ref": "#/components/schemas/MoneyResponse"
          },
          "ready_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "status": {
            "enum": [
              "CREATED",
              "PAYMENT_PENDING",
              "PAID",
              "GENERATING",
              "READY",
              "FAILED",
              "EXPIRED",
              "REFUNDED"
            ],
            "title": "Status",
            "type": "string"
          },
          "website_id": {
            "format": "uuid",
            "title": "Website Id",
            "type": "string"
          },
          "website_version_id": {
            "format": "uuid",
            "title": "Website Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "website_id",
          "website_version_id",
          "status",
          "price",
          "created_at",
          "paid_at",
          "ready_at",
          "expires_at"
        ],
        "title": "ExportPurchaseResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MoneyResponse": {
        "additionalProperties": false,
        "properties": {
          "amount_minor": {
            "minimum": 0.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "enum": [
              "INR",
              "USD"
            ],
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "MoneyResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Zylora V2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/export-prices": {
      "get": {
        "operationId": "export_prices_api_v1_admin_export_prices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExportPriceResponse"
                  },
                  "title": "Response Export Prices Api V1 Admin Export Prices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Export Prices",
        "tags": [
          "admin-exports"
        ]
      },
      "post": {
        "operationId": "configure_export_price_api_v1_admin_export_prices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportPriceConfigureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPriceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configure Export Price",
        "tags": [
          "admin-exports"
        ]
      }
    },
...
    "/api/v1/websites/{website_id}/exports": {
      "post": {
        "operationId": "create_export_purchase_api_v1_websites__website_id__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "website_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Website Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPurchaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Export Purchase",
        "tags": [
          "commerce"
        ]
      }
    },
...
  }
}